Redirect calls to printf and fprintf
When the interpreted code calls an external function, LLVM will use
dlsym() to find the target in the current process. This works well
for functions from libc which are always loaded.
However this also outputs the result of printf() and fprintf() to the
terminal and does not forward it to the user. This is very unintuitive
because values piped into std::cout and std::cerr correctly appear in
the browser.

As a solution, use llvm::sys::DynamicLibrary::AddSymbol() to inject
custom implementations into the process of finding external functions.
Once hooked up, the implementations format the passed arguments into
a C++ string and forward the result to the C++ streams that correctly
forward to the user.

Fixes #112